Move deplete function to dedicated pool module; removed timed_deplete
